libamba/node_metadata: rustfmt

+++ crates/model/src/control_flow.rs
@@ -236,7 +236,7 @@ mod test {
 	fn node(i: u32) -> NodeMetadata {
 		NodeMetadata::State {
 			amba_state_id: i,
-			s2e_state_id: i,
+			s2e_state_id: i as i32,
 		}
 	}
